Attempting to move browser-storage to packages.
